% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{bedcount_reader}
\alias{bedcount_reader}
\title{Read bedcount files}
\usage{
bedcount_reader(bedcount, compression = NULL, chunksize = 10000)
}
\arguments{
\item{bedcount}{Path to bedcount file}

\item{compression}{Compression type (default: NULL)}

\item{chunksize}{Chunk size for reading (default: 10000)}
}
\value{
Data frame with bedcount data
}
\description{
bedcount_reader returns a data frame of the bedcount data
}
\examples{
\dontrun{
data <- bedcount_reader("example.bedcount")
}
}
