% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpf_utils.R
\name{get_data}
\alias{get_data}
\title{Get Data from Tabix Files}
\usage{
get_data(
  files,
  labels = NULL,
  data_columns = NULL,
  regions = NULL,
  join = "outer",
  preset = "bed"
)
}
\arguments{
\item{files}{Character vector of file paths}

\item{labels}{Character vector of labels for the files}

\item{data_columns}{List of column specifications for each file}

\item{regions}{Data frame with regions to query}

\item{join}{Character string specifying join type (default: "outer")}

\item{preset}{Character string specifying file format preset (default: "bed")}
}
\value{
A list of data frames, one for each region
}
\description{
Combines tabix-indexed genome position files for specified regions.
}
