% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpf_utils.R
\name{groupby}
\alias{groupby}
\title{Group Data by Metadata Factor}
\usage{
groupby(by, metadata = NULL, data = NULL)
}
\arguments{
\item{by}{Character string specifying the grouping variable}

\item{metadata}{Data frame containing metadata}

\item{data}{Data frame containing the data to be grouped}
}
\value{
A grouped data structure
}
\description{
Group merged GPF data frame by levels of a factor.
}
