% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{groupname}
\alias{groupname}
\title{Generate group filename}
\usage{
groupname(by = NULL, name = NULL, fname = NULL)
}
\arguments{
\item{by}{Grouping variables}

\item{name}{Group name}

\item{fname}{Original filename}
}
\value{
Modified filename
}
\description{
Return filename of the subgroup based on grouping criteria
}
