% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{impute}
\alias{impute}
\title{Impute missing values}
\usage{
impute(data, method = "pseudocount")
}
\arguments{
\item{data}{Data frame with missing values}

\item{method}{Imputation method (default: 'pseudocount')}
}
\value{
Imputed value or modified data frame
}
\description{
Imputes missing values of a data frame using specified method
}
\examples{
\dontrun{
imputed_value <- impute(data, method = "pseudocount")
}
}
