% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{population_filter}
\alias{population_filter}
\title{Filter population data based on metadata}
\usage{
population_filter(metadata, subset = NULL, relation = NULL)
}
\arguments{
\item{metadata}{Path to metadata file}

\item{subset}{Subset condition (default: NULL)}

\item{relation}{Relation for grouping (default: NULL)}
}
\value{
List with reference and qset components
}
\description{
Read metadata and return the population and the quotient set
}
