% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{shannon_entropy}
\alias{shannon_entropy}
\title{Shannon Entropy Estimation}
\arguments{
\item{countmatrix}{A matrix of counts}

\item{axis}{Integer indicating which axis to compute entropy along (1 for rows, 2 for columns)}

\item{method}{Character string indicating the estimation method}
}
\value{
A vector of Shannon entropy values
}
\description{
Computes Shannon entropy (in nat) of the feature frequency profile using the plug-in estimator.
}
