% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpf_utils.R
\name{supremum_position}
\alias{supremum_position}
\title{Get Maximum Position}
\usage{
supremum_position(tabixfiles, chrom)
}
\arguments{
\item{tabixfiles}{Character vector of tabix-indexed file paths}

\item{chrom}{Character string specifying the chromosome}
}
\value{
Integer representing the maximum position, or NULL if no data
}
\description{
Return the least upper bound for the chromosome end coordinate.
}
