% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{js_divergence}
\alias{js_divergence}
\title{Jensen-Shannon Divergence}
\usage{
js_divergence(indata, debug = FALSE, compute_MET = TRUE)
}
\arguments{
\item{indata}{Input data frame.}

\item{debug}{Logical, whether to print debugging info. Default is FALSE.}

\item{compute_MET}{Logical, whether to compute MET column. Default is TRUE.}
}
\value{
A tibble with divergence values.
}
\description{
Calculates JSD from methylation count data.
}
